car_dealership/requirements.txt
Django
rest_framework_simplejwt
djangorestframework
boto3
django-storages
psycopg2-binary
django-cors-headers
environ

